First running Tests of Node-Python Communication
Added a Python Server that connects to a SocketIO in Node to facilitate
communications of gekko with python based indicators. Used the tulib and talib
indicator logic for easier implementation. First communications run
successfully and can be expanded upon.

TODO: check if what happens if no python server is running. Try the first
indicators that might take a while to compute and if the ticks are processed
subsequently and in the right order. Work out the python logic that can use
machine learning and neural nets for other data that send in the candle.