First running Prototype of Python-Indicators finished
Mostly work on the Python side of the interaction was done. A Python Module/
Package was created that handles that calculations of the Indicators and
hopefully facilitates the creation of new indicators. A first try of a test
LSTM indicator in a custom trading strategy was run.

TODO: Most of the machine learning methods require a custom scaling/detrending
of the date. This should somehow be directly incorporated in the model or
at least somehow be imported in the Indicator class. For now it is done "per
Hand"