Start of Revision of baseTradingMethod.js
First draft of a revised baseTradingMethod that is shorter and more concise.
It should be easier to understand the code now. Changes include the addition
of a tickQueue that every new tick is pushed on. The behavior of all
indicators has been unified. Additionally, the async indicators now return
a promise. This most likely doesn't work yet though.

TODO: Debug