Beta working Version of the PoloniexNN Indicator finished
Fixed the serialization of the scaler to use dill instead of pickle, which
makes it possible to export the scaler now. The Poloniex Data importer uses
multiple threads now, which enormly improves the speed of the Indicator. It
still takes a very long time. The first interactions with gekko are running
and backtests are now possible.

TODO: Fix the initialization from gekko. Maybe further improve speed somehow.
Improve and test the NN models.