Deleted the __pycache__ from git and added it to .gitignore
This folder does not need to be added, as it changes to much and there is
nothing to gain from having it on git. It will be created anyways when runnig.

Added the simple test for PoloniexNN and changed .gitignore

The test is a simplest case of a PoloniexNN and only goes through 2 dates.
.gitignore now contains python_indicators/__pycache__ and files in it.